Implemented canvas best practices
