+++ package.json
@@ -33,6 +33,7 @@
     "@types/node": "^22.1.0",
     "@types/react": "^18",
     "@types/react-dom": "^18",
+    "@types/three": "^0.170.0",
     "eslint": "^8.57.0",
     "eslint-config-next": "^14.2.5",
     "postcss": "^8",
